+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/ast/Java14PreviewTest.java
@@ -121,7 +121,7 @@ public class Java14PreviewTest {
 
         ASTRecordDeclaration range = recordDecls.get(2);
         Assert.assertEquals("Range", range.getSimpleName());
-        Assert.assertEquals(2, range.getRecordComponentList().size());
+        Assert.assertEquals(2, range.getRecordComponents().size());
         List<ASTRecordConstructorDeclaration> rangeConstructors = range.findDescendantsOfType(ASTRecordConstructorDeclaration.class);
         Assert.assertEquals(1, rangeConstructors.size());
         Assert.assertEquals("Range", rangeConstructors.get(0).getImage());
@@ -144,7 +144,7 @@ public class Java14PreviewTest {
 
         ASTRecordDeclaration emptyRec = recordDecls.get(5);
         Assert.assertEquals("EmptyRec", emptyRec.getSimpleName());
-        Assert.assertEquals(0, emptyRec.getRecordComponentList().size());
+        Assert.assertEquals(0, emptyRec.getRecordComponents().size());
 
         ASTRecordDeclaration personRec = recordDecls.get(6);
         Assert.assertEquals("PersonRecord", personRec.getSimpleName());
@@ -153,7 +153,7 @@ public class Java14PreviewTest {
     }
 
     private ASTRecordComponent getComponent(ASTRecordDeclaration arrayRec, int index) {
-        return arrayRec.getRecordComponentList().getChild(index);
+        return arrayRec.getRecordComponents().getChild(index);
     }